rename: MiniDialog as component
